Anchor TextField, Send Button and ScrollPane
These three components in GUI doesn't resize automatically.

No matter how the user resize the window, the components remain the
same size. This would greatly affect the overall user experience of
the GUI.

Set the components' anchor to the AnchorPane. Give the components
constraint to the AnchorPane's side.

Can be done easily through SceneBuilder and achieves the effect
accurately.